Start work on animations.
The plan is to have a low-level animations class and an animation collection
class that can accumulate them and play them together.  Every animation has
frames; every frame has what to draw, where to draw it, and when to draw it.

Once I've got a basic player working, the next step will be to make some
"tweening" utility functions in order to create animations that go from A to
be in X seconds, _hopefully_ smoothly transitioning their colors as they do
so.