Wrote a string tokenizer for Equations.
In order to convert strings to polynomials, I need a scanner, a parser, and an
evaluator.  The scanner I have now is pretty much good enough; the next step
is to write a parser that can convert the token string to an abstract syntax
tree (really, a postfix stack) using Edsger Dijkstra's Shunting Yard
Algorithm.