The Shunting Yard Algorithm is implemented, and it appears as though I now have a parser.
It worked the first time I wrote it!  All I had to add was the implement
multiplication stuff.  I based my implementation on the pseudocode at
https://en.wikipedia.org/wiki/Shunting-yard_algorithm#The_algorithm_in_detail,
but credit originally goes to Dijkstra.  That guy knew what he was doing!

The operators are in the correct order now for simple postfix evaluation, so
hopefully turning this token sequence into a polynomial will be
straightforward.

Oh.  And I have to handle unary operators somehow.