Added support for unary operators.
I followed the straightforward guidelines on
https://stackoverflow.com/questions/5239715/problems-with-a-shunting-yard-algorithm/5240781#5240781.
The parser now has support for not only unary operators of arbitrary
description, but right-associative operators in general.

It's time to start work on the evaluator.  It would need a host of algebraic
simplification rules to be truly general-purpose (for instance, how do you
evaluate NUMBER(5) SYMBOL(x) NUMBER(2) ADD MULTIPLY -- that is, "5(x + 2)"?)
For now, I'll stick with polynomial conversion and basic polynomial
arithmetic (with the exception of polynomial division, at least until the
synthetic division part is finished.)