Polynomials can now be added.
This was harder to implement than it looked due to the automatic merging of
compatible terms that needed to happen behind the scenes.

I can subtract by adding -1 * a polynomial, so the next operation to implement
is polynomial multiplication using FOIL (or something like it.)  I should also
add the ability to convert individual Terms to Polynomials just so I can print
them in a sensible way during debugging.