Refactored the colorstring-parsing code a bit.
For the forms that I'm planning to add, I want the ability to chop off the
ends of colorStrings "cleanly" (i.e., without breaking color codes and
revealing them to the user.)  For that, I created a "Substr(colorString,
start, len)" function....

...before realizing that it did a lot of the same work that
colorString.Print() did!  So I tried to refactor the overlapping work--the
parsing of C-style escape sequences and my own color codes -- into a
new method called readNextCharacter(), then have Substr() use _that_.

This commit accomplishes that task, and shores it up with a series of unit
tests that enforce the examples I have in the documentation.  The next step is
to rewrite Print() to use readNextCharacter() beneath the hood.