Substr() now correctly adds current color information, no matter where the string is sliced.
I was concerned that when rendering the current cursor in the middle of an
equation box, the original colors before the color change at the cursor would
not be restored because nothing saved that information.

I surmised that it should be Substr()'s job to copy color information just in
case we need it -- it's always harmless to add this to substrings as long as
the information is correct.