Just ran "go fmt."  Instantly regretted it.  Committing anyway.
The alignment of symbols for struct declarations is welcome; it makkes code
look better.  The elimination of whitespace around operators, on the other
hand, makes the code look *worse*.  And as Go developers, we're just supposed
to grin and bear it, because you can have any color of Model T as long as it's
black.

Ugh.  Still don't like this programming language.